requirements.txt
torch==1.10.1
torchvision==0.11.2
scikit-learn==0.24.2
pandas==1.2.4
seaborn==0.11.2
tqdm==4.61.1
tensorboard==2.9.1
matplotlib==3.5.2
joblib==1.1.0
numpy==1.20.3
requests==2.27.1
Pillow==9.0.1
scipy==1.8.0
